[C++, C#] Avoid unnecessary validation code.
Previously, regardless of whether the min and max values were the min
and max values for the host language representation of an SBE type, we
would generate a range check. Now, we do not generate the check unless
it is meaningful.

Also, adjust the property-based test dotnet version to bring inline with
other tests.